serial: 8250: Add special initialization for Keystone 2 UART
The UART IP within TI's Keystone 2 SOCs are 16550 compatible. However,
Keystone 2 devices also have a couple additional nonstandard UART
registers.

The UART Power Emulation and Management register must be written to for the
particular UART to be enabled. UART0 doesn't have this issue since this the
ROM or U-boot already handles this step. For other UARTs, writing to this
register is needed by the driver otherwise it won't work.

Create a new driver that will export a function that will be used by the
8250 core driver.

Signed-off-by: Sekhar Nori <nsekhar@ti.com>
[fcooper@ti.com: Slightly moved function calls around]
Signed-off-by: Franklin S Cooper Jr <fcooper@ti.com>