drm/omap: fix crtc properties
An error was made in the patch that adds the properties for omap_crtc when
it was forward ported to v4.4, causing crtc properties not to work at
all.

The problem was that all the new values for properties were set to the
old state, not the new state, so all new values just disappeared.

Signed-off-by: Tomi Valkeinen <tomi.valkeinen@ti.com>