net: prueth: Always inform the PRUs of PHY changes
Currently if the PHY state change handler detects that the PHY state has
not changed we do not update the PHY status in the PRU memory. This state
change handler checks the new state against a saved version of the state
stored in host memory, but does not verify that the PRU's knowledge of
the current state has not changed. The PRU's knowledge of the PHY state
is set in PRU DRAM that may be reset when the associated interface state
is changed. This can lead to the PRUs internal PHY state becoming out of
sync with the host's. Fix this here by always updating the PHY state in
the PRU's memory.

Fixes: 14d7cba3bbbb ("net: prueth: Add TI PRUSS Ethernet driver")
Reported-by: Aparna Balasubramanian <aparnab@ti.com>
Signed-off-by: Andrew F. Davis <afd@ti.com>
Tested-by: Aparna Balasubramanian <aparnab@ti.com>
Acked-by: Roger Quadros <rogerq@ti.com>
Signed-off-by: Sekhar Nori <nsekhar@ti.com>