mmc: host: omap_hsmmc: fixed tuning procedure for modes SDR104 and SDR50
The tuning procedure is mandatory for mode sdr104 and must be done for mode
sdr50 is the corresponding bit is set in the cap2 register of the card.
We previously relied on the clock frequency to select the kind of required tuning.
The consequence of this is that the tuning would not be done for a sdr104 card when
the bus frequency was set below 104 MHz.

This commit also fixes the path where the tuning is not required, by starting again
the mmc clock that had been stopped at the beginning of the procedure

Signed-off-by: Jean-Jacques Hiblot <jjhiblot@ti.com>