mmc: host: omap_hsmmc: lookup for pinctrl states based on the version tag
Silicon revision 1.1 and 1.0 of DRA7x silicon and silicon revision 2.0
comes with a different set of IOdelay numbers.
The driver can now handle multiple versions of IOdelay data through a
"version" tag to the pinctrl name.
This "version" tag is described in the DTS.

This commit an adapted subset of commit b68eb80264762 ("mmc: omap_hsmmc:
add support for pdata quirk") from ti-linux-4.1

Signed-off-by: Jean-Jacques Hiblot <jjhiblot@ti.com>