net/rpmsg: add support for new rpmsg sockets
Add the support for a new socket address and protocol
family: remote-processor sockets. The rpmsg protocol
driver provides the necessary support to expose a
rpmsg communication channel through the socket API
to userspace.

This patch is created from the initial rpmsg socket
support patch by Ohad on an older kernel.

Signed-off-by: Ohad Ben Cohen <ohad@wizery.com>
[rtivy@ti.com: forward port to 3.8 and add .bind functionality]
Signed-off-by: Robert Tivy <rtivy@ti.com>
[s-anna@ti.com: rproc id retrieval logic, cleanup and uapi header split,
		fixups for 4.1/4.4 kernels]
Signed-off-by: Suman Anna <s-anna@ti.com>