ARM: dts: dra72-evm-common: move mboxes and timers for rproc devices
The mailboxes and timer configuration will be the same for the IPUs
and the DSP1 remote processor for all the DRA72 EVM board variations
including the newer rev C board. So, move these data from the DRA72
EVM board dts file into the common dra72-evm-common dts file. The
reserved memory nodes and the enabling of the actual remoteproc
devices still needs to be managed within the respective board dts
file.

Signed-off-by: Suman Anna <s-anna@ti.com>