ARM: dts: keystone-k2e: Add PSC reset control data
The Power Sleep Controller (PSC) module contains specific
memory-mapped registers that can be used to perform reset
management for the DSPs available on the SoC using specific
bits. The PSC is defined using a syscon node, and the reset
functionality is defined using a child syscon reset controller
node. Add the reset control data to this syscon reset controller
node for the Keystone 2 Edison SoC which the single consumer
DSP device can refer to for managing its reset.

Signed-off-by: Andrew F. Davis <afd@ti.com>
Signed-off-by: Suman Anna <s-anna@ti.com>
Signed-off-by: Dave Gerlach <d-gerlach@ti.com>