ARM: dts: am57xx-beagle-x15-revb1: fix HDMI LS_OE gpio
Commit 86cd2e41104261d3af1a3dfc89d25848f5dd2f7a ("ARM: dts:
am57xx-beagle-x15: Add support for rev B1") adds HDMI changes needed for
new x15 boards. However, while the description of the patch is correct,
the change itself is not correct. In fact, it's just a copy of the gpios
already in the am57xx-beagle-x15-common.dtsi files.

This patch fixes the LS_OE gpio.

Signed-off-by: Tomi Valkeinen <tomi.valkeinen@ti.com>