ARM: dts: dra7-evm.dts: Limit the speed of MMC1 and MMC2
Per erattum i843 the clocks of the mmc1 and mmc2 controllers must be
limited respectively to 96MHz and 48MHz on rev 1.1 and earlier.
On those platforms, using mmc1 or mmc2 at full speed (192MHz) will result
in data corruption. As the settings for rev 1.1 should also be okay for
rev 2.0 platforms, let's use the settings for rev 1.1 by default.

Signed-off-by: Jean-Jacques Hiblot <jjhiblot@ti.com>