DRA7: hsmmc: quirks: Adjust the max frequency based on SOC revision
The max frequency of the SOC revisions prior 2.0 are limited (erratum i843)
To allow a DTB to  fit any SOC revision, we need to tweak it at startup.
This patch reads the SOC revision and adjusts the max frequency of the
controller and sets the hw revsion in the DTS for the hsmmc driver.

Signed-off-by: Jean-Jacques Hiblot <jjhiblot@ti.com>