ARM: dts: dra7-evm.dts: Remove the max-frequency of the hsmmmc controllers
The maximum speed available for a hsmmc controller is checked at startup
by a pdata-quirk. If no maximum frequency is provided then the frequency is
set to its maximum value.
The pdata-quirk will also set the correct revision string for the platform.

So remove this static information from the DTS and let the kernel handle it
dynamically.

Signed-off-by: Jean-Jacques Hiblot <jjhiblot@ti.com>