HACK: ARM: dts: AM33xx: Split OCMC RAM node to allow non-cached mapping
Split the OCMC RAM node into two nodes of sizes 8 KB and 56 KB. The
smaller node continues to be marked with the 'map-exec' property for
making this SRAM portion executable and cacheable, this will be used
to achieve the AM335x Power Management functionality.

The larger node uses the default settings, allowing it to be mapped
using the default mapping scheme (write-combine, non-cached) by the
SRAM driver, and will allow general purpose usage including the
PRUSS Ethernet usecase on specific platforms.

The sizes can be finetuned, if needed, based on memory consumption
requirements for different usecases.

Signed-off-by: Suman Anna <s-anna@ti.com>
Tested-by: Dave Gerlach <d-gerlach@ti.com>