TEMP: remoteproc/pru: use different firmware names for ethernet
The PRUSS Ethernet usecase is currently designed to have the PRUSS
ethernet driver perform the load and boot of a PRU device (in the
expected fashion when a firmware doesn't support virtio devices),
though the ethernet driver wants the flexibility to use a specific
firmware name, and possibly supply the firmware name in the future.
The remoteproc core instrastructure currently does the loading as a
two-step process with the resource table processed and cached the
first step, and a CRC-based cross-check in the second step, with
the firmware name fixed at the remoteproc registration time. The
commit cba380afd953 ("HACK: remoteproc/pru: manage PRUs without
virtio devices manually") has also provided a temporary mechanism
to manage the boot cycle of PRU cores running firmwares without any
virtio devices previously for generic usecases.

It cannot be currently determined whether the PRUs are being used for
an Ethernet usecase or the previously supported generic usecases. A
decision is made by introducing a module parameter 'use_eth_fw' and
on the assumption that Ethernet usecases are to be supported only on
IDK boards. Even on current AM57xx IDK boards, PRUSS1 is not capable
of enabling Ethernet functionality due to board pin mux issues, so
augment the module parameter based decision making with an instance
specific functionality capabilities flag.

The above assumptions and/or solution is not solid and is not
scalable in general. Real fixes will need restructuring work done
on the remoteproc core to facilitate client provided firmware names
at run-time and not rely on firmware names provided at respective
platform driver probe time; and a rework on the PRUSS remoteproc
driver integration aspects with its client users to meet its myriad
usecases.

Signed-off-by: Andrew F. Davis <afd@ti.com>
Signed-off-by: Suman Anna <s-anna@ti.com>