ARM: dts: am57xx-idk-common: Add dcan support
Both of the AM571x-idk and the AM572x-idk use CAN1 interface.
This patch enables it for both boards.

Per this commit pin mux support is required:
commit 1fdfca820e34 ("ARM: dts: dra7-evm: add back DCAN1 pinmux")
Author: Sekhar Nori <nsekhar@ti.com>

Referenced commit message:
Muxing for DCAN pins is still needed in kernel to workaround the
issue of DCAN module failing to idle (i.e. stuck in transition)
if the module was disabled while there was traffic on the CAN_RX line.

Signed-off-by: Schuyler Patton <spatton@ti.com>
Acked-by: Sekhar Nori <nsekhar@ti.com>
Signed-off-by: Kishon Vijay Abraham I <kishon@ti.com>
Signed-off-by: Franklin S Cooper Jr <fcooper@ti.com>
Signed-off-by: Sekhar Nori <nsekhar@ti.com>