drm/omap: check plane size
At the moment we don't check the plane input/output sizes, which can
lead to DSS HW errors when invalid values are given from the userspace.

Add a check so that the sizes are > 0.

Note: the feedback from mainline was that plane size of 0 means
"disabled". In the future this should be implemented that way.

Signed-off-by: Tomi Valkeinen <tomi.valkeinen@ti.com>