ARM: dts: am437x: add DT node for gpu
Add dt node to enable SGX PowerVR driver in AM437x base dtsi file.

Note: The SGX is disabled by default because it is an optional module
and only some AM437x variants contain this module, such as AM4378 and
AM4379.  The status should be overwritten as "OK" at the corresponding
board.dts.

Signed-off-by: Darren Etheridge <detheridge@ti.com>
Signed-off-by: Eric Ruei <e-ruei1@ti.com>