net: netcp: rearrange variable in struct for better cache line use
This patch identified struct netcp_intf field, rx_queue_depths,
used in the data path and move it ahead of the non used variable
for better cache line use. This patch also add a cautionary note
to help developers add new variables to the right place in netcp_intf
struct as it is used in data path and proper cache line usage should
be ensured.

Signed-off-by: Murali Karicheri <m-karicheri2@ti.com>
Signed-off-by: Sekhar Nori <nsekhar@ti.com>