ARM: dts: AM4372: Add the PRU-ICSS1 DT node
Add the DT node for the PRU-ICSS1 on AM437x family of SoCs.
The DT node uses all standard properties. The regs property
defines the addresses for various modules within the PRUSS,
including all the Data RAMs, the Instruction RAM per core,
the INTC, the Industrial Ethernet Peripheral (IEP), the
Real Time Media Independent Interface controller (MII_RT),
CFG, Debug and Control sub-modules.

The two PRU cores within the PRUSS and the interrupt controller
are represented as child nodes (so platform devices themselves)
of the PRUSS parent node.

The PRU-ICSS1 on AM437x is very similar to the PRUSS in AM33xx,
except for variations in the RAM sizes, bus addresses and the
number of interrupts coming into the MPU INTC (host interrupt
7 is routed to the other PRUSS instead of MPU).

Signed-off-by: Suman Anna <s-anna@ti.com>
Signed-off-by: Andrew F. Davis <afd@ti.com>