ti_config_fragments/baseport.cfg: Change amx3 PM code to modules
In order to prepare for systemd migration and also because there is no
reason not to, change the drivers needed for amx3 PM to build as
modules.

Both udev and eudev have removed the userspace firmware loader helper,
so building these as modules will allow them to be loaded at the
same time the PM firmware is available.

Signed-off-by: Dave Gerlach <d-gerlach@ti.com>
Acked-by: Suman Anna <s-anna@ti.com>