remoteproc/pruss: enable OCP master ports in SYSCFG always
The PRUSS CFG module's SYSCFG register is used for managing the
PRCM clock management settings at the PRU-ICSS subsystem level.

Add two helper functions pruss_{enable/disable}_module() that
programs this SYSCFG register. The register is currently programmed
to enable the OCP master ports (disable MStandby), and for the
default Smart-Idle and Smart-Standby always. This is done to
allow the PRUs access to on-chip memories and peripherals outside
the PRUSS without having to do it in firmware (primarily for the
PRU Ethernet usecase currently, the firmware is not programming
this register).

NOTE:
1. The AM57xx TRM suggests to enable "No Standby" during the
   Idle + Standby state to an Active/Normal state, but this
   sequence is not documented in either of AM33xx and AM437x
   SoCs. Furthermore, it did not have an impact on the working
   of Ethernet Rx, so is left out.
2. This has an impact on the PM suspend/resume operation usually
   (same bit is used to trigger PRU standby), and even when the
   PRU cores are in halted state. The STANDBY_INIT has to be
   re-programmed to initiate a standby sequence and have the PM
   suspend/resume functional when PRUs are halted. This is already
   handled in commit 12cb6cb49bca ("remoteproc/pruss: fix system
   suspend/MStandby config issues").

Signed-off-by: Roger Quadros <rogerq@ti.com>
Signed-off-by: Andrew F. Davis <afd@ti.com>
Signed-off-by: Suman Anna <s-anna@ti.com>