media: ti-vpe: vip: Improve scaler allocation to only when needed
The original scaler support was allocating the scaler module to
the first open port of a VIP slice. Similarly it was only release
when the port was closed/released.
So even if the requested frame size was supported natively by the
sensor the scaler module was still assigned to the port but in bypass
mode.

This patch expand on this by creating a scaler resource allocation
mechanism and only attempt to allocate the scaler unit if the
requested frame size is not already natively supported by the sub
device.

So instead of being allocated always on device open, it will only be
allocated if needed during a S_FMT ioctl. And freed either when the
device node is closed or on subsequent S_FMT ioctl call if no longer
needed.

Signed-off-by: Benoit Parrot <bparrot@ti.com>