drm/omap: tc358768: search for best PLL output rate
The driver tries to find a config for the PLL which produces an exact
match for the requested pixel clock. This often doesn't work, so we need
to have a bit of a tolerance there.

This patch makes the driver search for a best match, instead of an exact
match.

Signed-off-by: Tomi Valkeinen <tomi.valkeinen@ti.com>