drm/omap: Use omapdss_stack_is_ready() to check that the display stack is up
Instead of 'guessing' based on aliases of the status of the DSS drivers,
use the new interface to check that all needed drivers are loaded.
In this way we can be sure that all needed drivers are loaded so it is
safe to continue the probing of omapdrm.
This method will allow the omapdrm to be probed 'headless', without
outputs.

Fixes: 3ae08caa8c1ba (drm/omap: ensure all displays have been probed)
Signed-off-by: Peter Ujfalusi <peter.ujfalusi@ti.com>