ARM: dts: dra72-evm: fix iodelay configuration for mmc2 HS200 mode
The manual mode timings configured for MMC2 HS200 mode
do not match the latest data manual for AM571x device
SPRS919K, revised February 2016[1].

Fix this and bring configuration back in sync with the
latest data manual. While at it, also rearrange the
configuration entries in the same order as they appear
in data manual so it is easy to cross-verify their
correctness.

[1] http://www.ti.com/lit/ds/symlink/am5718.pdf

Signed-off-by: Sekhar Nori <nsekhar@ti.com>