ARM: dts: am437x-idk-evm: Add reset to PRUSS1 ethernet PHYs
The PRUSS1 ethernet PHYs are in invalid state during system
reset due to incorrect pin states of the MII bus.

We need to reset the PHY after the PHY pins have been brought to
the right state to make sure the PHY is in a good state
else it causes unreliability issues like PHY link not working
or PHY not being detected on MDIO bus.

Signed-off-by: Andrew F. Davis <afd@ti.com>
Signed-off-by: Sekhar Nori <nsekhar@ti.com>