rpmsg: add helper macro module_rpmsg_driver
[ Upstream commit f3d9f1ce079370d19b34fc9927b8b1355ac98503 ]

This patch introduces the module_rpmsg_driver macro which is a
convenience macro for rpmsg driver modules similar to
module_platform_driver. It is intended to be used by drivers which
init/exit section does nothing but register/unregister the rpmsg driver.
By using this macro it is possible to eliminate a few lines of
boilerplate code per rpmsg driver.

Signed-off-by: Andrew F. Davis <afd@ti.com>
Signed-off-by: Bjorn Andersson <bjorn.andersson@linaro.org>
[s-anna@ti.com: cherry-pick commit 'f3d9f1ce0793' from v4.7]
Signed-off-by: Suman Anna <s-anna@ti.com>