Documentation: dt: add bindings for ti-cpufreq
Add the device tree bindings document for the TI CPUFreq/OPP driver
on AM33xx and AM43xx SoCs. This driver is responsible reading and
parsing registers that define which OPPs are available for CPUFreq to
use. From this is sets supported-hw for the OPP framework so that OPPs
can be selectively enabled based on the specific SoC in use.

Signed-off-by: Dave Gerlach <d-gerlach@ti.com>