ARM: dts: keystone-k2g: Add the PRU-ICSS nodes
Add the DT nodes for the PRU-ICSS0 and PRU-ICSS1 processor
subsystems that are present on the K2G SoC. The two PRU-ICSSs
are identical to each other.

The DT nodes uses all standard properties. The regs property
defines the addresses for various modules within the PRUSS,
including all the Data RAMs, the Instruction RAM per core,
the INTC, the Industrial Ethernet Peripheral (IEP), the
Real Time Media Independent Interface controller (MII_RT),
CFG, Debug and Control sub-modules. The two PRU cores within
the PRUSS and the interrupt controller are represented as
child nodes (so platform devices themselves) of the PRUSS
parent node.

The PRU-ICSSs on K2G are very similar to the PRUSS on the
AM57xx SoCs except for larger Shared Data RAM and the lack
of a PRU-ICSS crossbar. There are also few other minor
integration differences w.r.t IPC mechansims that can be
attributed to the architecture differences between Keystone
and OMAP families.

Signed-off-by: Andrew F. Davis <afd@ti.com>
[s-anna@ti.com: intc adjustment, added commit message]
Signed-off-by: Suman Anna <s-anna@ti.com>