drm/omap: use print instead of WARN when plane setup fails
We do WARN() when dispc's ovl_set() fails. This should really not
happen in normal operation, as omapdrm should verify in atomic_check
that the plane setup is correct. However, we don't check much there, and
e.g. any scaling config errors are not detected.

Instead of spamming the log with WARNs when the user gives a bad scaling
config, just print a single line.

Just is just a quick fix, as we should really do a much more thorough
check in atomic_check.

Signed-off-by: Tomi Valkeinen <tomi.valkeinen@ti.com>