remoteproc/pruss: localize INTC functions to pruss-intc module
The PRUSS INTC management is currently split between the pruss module
and the pruss-intc modules. The pruss module currently provides two
API for the pru-rproc driver to map the PRUSS system events into
interrupt channels and host interrupts based on data parsed from
PRU remoteproc firmware images. The pruss_intc module implements
an irqchip driver to provide users standard API to enable/disable
interrupts based on these system events. The two API and related
variables are moved from the pruss module to the pruss-intc module
so that all interrupt management is localized to a single module.

NOTE:
The INTC address space is still parsed in pruss module, this will
be localized to pruss-intc node and module in a follow-up patch.

Signed-off-by: Suman Anna <s-anna@ti.com>