ARM: dts: dra7: Fix the minimum and maximum Voltage for CPU OPPs
OPP table as defined for DRA7/AM57xx Datasheets define the voltage
ranges possible for AVS class0 voltage, however, the % accuracy
permitted is NOT the voltage tolerance parameter we use for the PMIC, it
is the % tolerance allowed by the PMIC for a determined voltage value at
the SoC ball level. Fix the data as per the required documentation.

The exception is OPP_OD voltage range which we use the historical data
to ensure that SoCs from older lots still continue to boot, even though
more optimal voltages is now the standard.

Data based on: SPRUHQ0H Nov 2015

Reported-by: Carlos Hernandez <ceh@ti.com>
Signed-off-by: Nishanth Menon <nm@ti.com>
Acked-by: Dave Gerlach <d-gerlach@ti.com>