ti_config_fragments: baseport: Enable CONFIG_DEBUG_INFO by default
Build kernel with debug information so that folks can use the vmlinux
generated with JTAG and be able to live debug of the code. This costs us
almost nothing to enable, but enables folks who do crash analysis from
product image testing and developers trying to reproduce the errors from
having to rebuild kernel to figure out the bug.

This adds to vmlinux:
add/remove: 0/0 grow/shrink: 1/0 up/down: 22/0 (22)

Reported-by: Grygorii Strashko <grygorii.strashko@ti.com>
Signed-off-by: Nishanth Menon <nm@ti.com>