media: ti-vpe: vip: Validate frame size based on pixel format
commit 28d8b4290e70ff3c8538dc1b4847372e4b653437 ti-linux-4.1.y

Up to now VIP would allow frame size based on the sensor supported list.
However depending on the target pixel format some size would produce
unaligned vpdma size (16 bytes alignment required).

For instance 600x400 in NV12 is not 16 bytes aligned as 600 is not
divisible by 16. Leaving it as is would produce frame that would be
608 bytes wide.

According to v4l2-compliance tools driver are responsible to provide
a valid frame size and format. In doing so we cannot just fail in
this scenario but find a suitable value and use that instead.

This patch will attempt to find a valid frame size as close as possible
to the one asked for.

Signed-off-by: Benoit Parrot <bparrot@ti.com>
Signed-off-by: Jyri Sarha <jsarha@ti.com>
Signed-off-by: Jean-Jacques Hiblot <jjhiblot@ti.com>

Conflicts:
	drivers/media/platform/ti-vpe/vip.c