ARM: dts: am57xx-beagle-x15: Replace mailboxes with PRU system events
The virtio/rpmsg communication between MPU and a PRU core is currently
achieved using the OMAP mailbox IP that provides the inter-processor
interrupt signalling. Switch these mailboxes to use the PRU system
events so that the interrupt signalling can be achieved through a
mechanism internal to the PRUSS itself, and eliminating an external
peripheral access from the PRU-side. The difference from MPU would
be minimal in using one versus the other.

This switch is done for all AM57xx GP EVM and BeagleBoard-X15 boards.
Mailboxes can still be used if desired. Either approach would require
that an appropriate firmware image is loaded/booted on the PRU.

Signed-off-by: Andrew F. Davis <afd@ti.com>
Signed-off-by: Suman Anna <s-anna@ti.com>