HACK: net: prueth: Fix net transmit queue stall
When the PRU packet transmit buffer fills we stop the kernel network
queue for this interface, then when the PRU signals us that a packet has
been transmitted we re-enable this queue so we can again receive packets
and check again for enough buffer space. There currently is a situation
when under high network traffic we do not re-enable the queue, possibly
due to a race condition and/or firmware issue. The temporary fix is to
not disable the queue at all.

Fixes: 67c183326fcc ("net: prueth: Add TI PRUSS Ethernet driver")

Reported-by: Aparna Balasubramanian <aparnab@ti.com>
Signed-off-by: Andrew F. Davis <afd@ti.com>
Acked-by: Roger Quadros <rogerq@ti.com>