ti_config_fragments/defconfig_map.txt: Get rid of defconfig_fragment
We describe the base defconfig, the config fragments in the correct
order in the map file, this makes the defconfig_fragment redundant.

Replace it in map file with "None." (or unused) and replace with the
contents as extra configs

NOTE: only the TISDK defaults are updated here.

While at this do the following:
- Adjust the order of the cfg override such that SoC override is applied
just after the family override for omap architecture.
- Introduce lpae.cfg which is needed for larger memories on DRA7
platforms.

Signed-off-by: Nishanth Menon <nm@ti.com>