ti_config_fragments/defconfig_map.txt: Get rid of multi_core_defconfig_fragment
We describe the base defconfig, the config fragments in the correct
order in the map file, this makes the defconfig_fragment redundant.

Replace it in map file with "None." (or unused) and replace with the
contents as extra configs

NOTE: only the TISDK defaults are updated here.

While at it, introduce the lpae.cfg in the right order.

Signed-off-by: Nishanth Menon <nm@ti.com>