ti_config_fragments/defconfig_map.txt: Fix ordering of cfg files
cfg files need to be organized in a dependable manner. The sane option
is as follows:
defconfig -> prune -> Arch -> SoC specific(optional) -> LPAE
(optional)-> baseport -> ipc -> connectivity -> audio_display.

The rationale is as follows: as we prune the defconfig in each stage,
the next level of prune/enable can now depend on the previous stage.

Further, as we introduce RT/Systemd/Android/ or other variants, the
defconfig fragment cfg files related to these can be introduced in a
consistent manner.

NOTE that ipc.cfg introduction is done early for PRUSS/messaging
dependencies to be available.

Signed-off-by: Nishanth Menon <nm@ti.com>