cpufreq: ti-cpufreq: Assign defaults to max frequency efuse values
Some of the SoCs have the efuse registers which determine the max
frequency of cpu reading 0. This in turn gets translated to all
OPPs enabled and some of the OPPs which have matching frequencies
will error out as duplicate OPPs and cpufreq will be not functional.

The problem is with AM335X PG2.0 silicon whose max frequency efuse values
are not populated. This is true for all 2.0 silicon on Beaglebone
black and Beaglebone.

Hence assigning default efuse values for SoCs which read 0.

Signed-off-by: Keerthy <j-keerthy@ti.com>
Acked-by: Dave Gerlach <d-gerlach@ti.com>