phy: ti-pipe3: Program the DPLL even if it was already locked
If bootloader has set a wrong DPLL then we must trash those values
and re-program it anyways. This fixes USB3 devices not being enumerated
on beagle-x15 if usb was started in u-boot. (e.g. v2016.05)

We don't re-program SATA DPLL if it is locked as it was causing
SATA failures if device was hotpluged after boot. This might be due
to i818 "SATA PHY Reset Required Following SATA PLL Unlock".

Signed-off-by: Roger Quadros <rogerq@ti.com>
Signed-off-by: Sekhar Nori <nsekhar@ti.com>