remoteproc/pruss: fix NULL pointer de-reference
A PRU rproc device can be present but be unregistered with the
remoteproc core if the driver probe bailed out due to some failure
(e.g. firmware not present on FS resulting in auto-boot failure).
The pruss_rproc_get() can result in a NULL pointer dereference
due to the missing error check on the platform_get_drvdata() in
_pruss_rproc_get() function. Fix this by checking for non-NULL
rproc pointer.

Fixes: 7f25e1b27108 ("remoteproc/pruss: add APIs to get and set the PRU cores")
Signed-off-by: Roger Quadros <rogerq@ti.com>
Signed-off-by: Suman Anna <s-anna@ti.com>