ARM: dts: keystone: remove bogus PCI IO resources in DT bindings
In the past, the PCI controller required a IO resource to be specified
without which it fails in PCI core when creating the RC device. So
to overcome this, a bogus entry was added. In recent kernel, this
requirement is removed and we could initialize without specifying an IO
resource. Keystone PCI controller hardware doesn't support IO window.
So it is not required.

The flip side of this is that during boot, following error log is thrown
in, but doesn't cause any issues with SATA hard drive access. So remove
the bogus entry as having it can causes the device to fail for access to
the IO window (only snapshot of first few lines of dump shown)

[   37.182381] Unable to handle kernel paging request at virtual address fee01020
[   37.189598] pgd = d9d5fec0
[   37.192297] [fee01020] *pgd=80000800007003, *pmd=00000000
[   37.197700] Internal error: Oops: 206 [#1] PREEMPT SMP ARM
[   37.203176] Modules linked in: xhci_plat_hcd ahci libahci dwc3 libata udc_core extcon dwc3_keystone
[   37.212258] CPU: 0 PID: 1797 Comm: udevadm Not tainted 4.4.11-00323-ge3a9adb #1
[   37.219557] Hardware name: Keystone
[   37.223038] task: daaec600 ti: d9eba000 task.ti: d9eba000
[   37.228433] PC is at pci_read_resource_io+0xe4/0xf4
[   37.233302] LR is at 0x0
[   37.235830] pc : [<c02e32f8>]    lr : [<00000000>]    psr: 60010013
[   37.235830] sp : d9ebbe50  ip : da11a240  fp : d9ebbe6c
[   37.247295] r10: d9fe4000  r9 : da31e9cc  r8 : d9ebbf80
[   37.252511] r7 : 00000000  r6 : 00001023  r5 : 00000000  r4 : 00001023
[   37.259029] r3 : d9fe4000  r2 : fee01020  r1 : 00001020  r0 : 00000004
[   37.265548] Flags: nZCv  IRQs on  FIQs on  Mode SVC_32  ISA ARM  Segment user
[   37.272673] Control: 30c5387d  Table: 19d5fec0  DAC: fffffffd
[   37.278410] Process udevadm (pid: 1797, stack limit = 0xd9eba210)

Signed-off-by: Murali Karicheri <m-karicheri2@ti.com>
Signed-off-by: Sekhar Nori <nsekhar@ti.com>