rtc: ds1307: Fix the weekday register value
The reset value of weekday is 0x1. This is wrong since
the reset values of the day/month/year make up to Jan 1 2001.
When computed weekday comes out to be Monday. On a scale
of 1-7(Sunday - Saturday) it should be 0x2. So we should not
be relying on the reset value.

Hence compute the wday using the current time values.
Check if reset wday is any different from the computed wday
If different then set the wday which we computed using timestamp.

Document Referred:
http://ww1.microchip.com/downloads/en/DeviceDoc/20002266F.pdf

Signed-off-by: Keerthy <j-keerthy@ti.com>