drm/omap: Add top level WB support driver
In preparation to add WB Capture mode we need to separate the
top level functions used to register with omapdrm and install
irq hooks.

As we are splitting the omapdrm registration helpers
to a new files we also need to update omap_drv to
use the more generic WB hooks as well as the
related Kconfig and Makefile changes.

Signed-off-by: Benoit Parrot <bparrot@ti.com>