remoteproc/pruss_intc: dispose irq_mappings during remove
The PRUSS INTC driver provides an irqchip interface for user clients
to use one of the 64 PRU system events as interrupt sources. The irq
descriptors for these are usually allocated at the time of a user
device probe or during the IRQ resource allocation of a user device
creation from DT. These irq mappings are not currently released when
the PRUSS INTC driver is removed or INTC device is unbound. A new set
of mappings are created when the driver is reprobed again, and this
results in stale irq descriptors.

Fix this by disposing the irq mappings in the driver remove. The irq
data is also reset through the irq_domain's unmap ops. This eliminates
any kernel crashes due to the usage of any stale irq_mapping data by
user clients.

Fixes: 5b107622d1eb ("remoteproc/pruss: add platform drivers for PRUSS in TI SoCs")
Signed-off-by: Suman Anna <s-anna@ti.com>